+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Fix return type of `FileSystemCache::load` to `VmResult<Option<Module>>` in
   order to differentiate missing files from errors.
 - Add in-memory caching for recently used Wasm modules.
+- Rename `CosmCache` to just `cosmwasm_vm::Cache`.
 
 ## 0.11.2 (2020-10-26)